Extract the diagnostics implementation into its own module
check:architecture caps lifecycle.rs at 1000 lines and the diagnostics
implementation pushed it to 1095. Moving it to provider/diagnostics.rs keeps
the budget meaningful rather than raising it, and the sections belong together
anyway. Also applies cargo fmt.
